fix(suite): Fix styling and simple refactoring
