+++ packages/components/src/index.ts
@@ -13,6 +13,7 @@ export * from './components/Flag/Flag';
 export * from './components/AutoScalingInput/AutoScalingInput';
 export * from './components/Badge/Badge';
 export * from './components/buttons/Button/Button';
+export { type ButtonVariant } from './components/buttons/buttonStyleUtils';
 export * from './components/buttons/ButtonGroup/ButtonGroup';
 export * from './components/buttons/IconButton/IconButton';
 export * from './components/Icon/Icon';
