+++ packages/suite/src/components/suite/layouts/SuiteLayout/PageHeader/GlobalSendReceive/GlobalSendModal.tsx
@@ -1,0 +1,21 @@
+import { Account } from '@suite-common/wallet-types';
+
+import { GlobalSendReceiveModalBase } from './GlobalSendReceiveModalBase';
+import { LocalAccountSearchProvider } from '../../../../../../hooks/suite/useAccountSearch';
+import { AccountItemType } from '../../../../../../types/wallet';
+import { Translation } from '../../../../Translation';
+
+type GlobalSendModalProps = {
+    onCancel: () => void;
+    onSubmit: (account: Account, type: AccountItemType) => void;
+};
+
+export const GlobalSendModal = ({ onCancel, onSubmit }: GlobalSendModalProps) => (
+    <LocalAccountSearchProvider>
+        <GlobalSendReceiveModalBase
+            heading={<Translation id="SEND_TRANSACTION" />}
+            onCancel={onCancel}
+            onSubmit={onSubmit}
+        />
+    </LocalAccountSearchProvider>
+);
